Make array of floats to hold more bits

+++ Lab3/Project1/main.c
@@ -1,6 +1,6 @@
 #include <msp430.h> 
 
-unsigned long fib [100];
+float fib [100];
 
 /*
  * main.c
